+++ mayan/apps/documents/tests/test_trashed_document_models.py
@@ -1,4 +1,4 @@
-from ..models import DeletedDocument, Document
+from ..models import TrashedDocument, Document
 
 from .base import GenericDocumentTestCase
 
@@ -9,12 +9,12 @@ class TrashedDocumentTestCase(GenericDocumentTestCase):
 
         # Trash the document
         self.test_document.delete()
-        self.assertEqual(DeletedDocument.objects.count(), 1)
+        self.assertEqual(TrashedDocument.objects.count(), 1)
         self.assertEqual(Document.valid.count(), 0)
 
         # Restore the document
         self.test_document.restore()
-        self.assertEqual(DeletedDocument.objects.count(), 0)
+        self.assertEqual(TrashedDocument.objects.count(), 0)
         self.assertEqual(Document.valid.count(), 1)
 
     def test_trashing_documents(self):
@@ -22,10 +22,10 @@ class TrashedDocumentTestCase(GenericDocumentTestCase):
 
         # Trash the document
         self.test_document.delete()
-        self.assertEqual(DeletedDocument.objects.count(), 1)
+        self.assertEqual(TrashedDocument.objects.count(), 1)
         self.assertEqual(Document.valid.count(), 0)
 
         # Delete the document
         self.test_document.delete()
-        self.assertEqual(DeletedDocument.objects.count(), 0)
+        self.assertEqual(TrashedDocument.objects.count(), 0)
         self.assertEqual(Document.valid.count(), 0)