+++ mayan/apps/documents/tests/test_trashed_document_views.py
@@ -1,4 +1,4 @@
-from ..models import DeletedDocument, Document
+from ..models import TrashedDocument, Document
 from ..permissions import (
     permission_trashed_document_delete, permission_trashed_document_restore,
     permission_document_trash, permission_document_view,
@@ -36,7 +36,7 @@ class TrashedDocumentViewTestCase(
         response = self._request_test_document_trash_post_view()
         self.assertEqual(response.status_code, 404)
 
-        self.assertEqual(DeletedDocument.objects.count(), 0)
+        self.assertEqual(TrashedDocument.objects.count(), 0)
         self.assertEqual(Document.valid.count(), 1)
 
     def test_document_trash_post_view_with_access(self):
@@ -47,7 +47,7 @@ class TrashedDocumentViewTestCase(
         response = self._request_test_document_trash_post_view()
         self.assertEqual(response.status_code, 302)
 
-        self.assertEqual(DeletedDocument.objects.count(), 1)
+        self.assertEqual(TrashedDocument.objects.count(), 1)
         self.assertEqual(Document.valid.count(), 0)
 
     def test_trashed_document_restore_get_view_no_permission(self):
@@ -83,7 +83,7 @@ class TrashedDocumentViewTestCase(
         response = self._request_test_trashed_document_restore_post_view()
         self.assertEqual(response.status_code, 404)
 
-        self.assertEqual(DeletedDocument.objects.count(), 1)
+        self.assertEqual(TrashedDocument.objects.count(), 1)
         self.assertEqual(Document.valid.count(), 0)
 
     def test_trashed_document_restore_post_view_with_access(self):
@@ -97,56 +97,56 @@ class TrashedDocumentViewTestCase(
         response = self._request_test_trashed_document_restore_post_view()
         self.assertEqual(response.status_code, 302)
 
-        self.assertEqual(DeletedDocument.objects.count(), 0)
+        self.assertEqual(TrashedDocument.objects.count(), 0)
         self.assertEqual(Document.valid.count(), 1)
 
     def test_trashed_document_delete_get_view_no_permission(self):
         self.test_document.delete()
         self.assertEqual(Document.valid.count(), 0)
-        self.assertEqual(DeletedDocument.objects.count(), 1)
+        self.assertEqual(TrashedDocument.objects.count(), 1)
 
-        trashed_document_count = DeletedDocument.objects.count()
+        trashed_document_count = TrashedDocument.objects.count()
 
         response = self._request_test_trashed_document_delete_get_view()
         self.assertEqual(response.status_code, 404)
 
         self.assertEqual(
-            DeletedDocument.objects.count(), trashed_document_count
+            TrashedDocument.objects.count(), trashed_document_count
         )
 
     def test_trashed_document_delete_get_view_with_access(self):
         self.test_document.delete()
         self.assertEqual(Document.valid.count(), 0)
-        self.assertEqual(DeletedDocument.objects.count(), 1)
+        self.assertEqual(TrashedDocument.objects.count(), 1)
 
         self.grant_access(
             obj=self.test_document, permission=permission_trashed_document_delete
         )
 
-        trashed_document_count = DeletedDocument.objects.count()
+        trashed_document_count = TrashedDocument.objects.count()
 
         response = self._request_test_trashed_document_delete_get_view()
         self.assertEqual(response.status_code, 200)
 
         self.assertEqual(
-            DeletedDocument.objects.count(), trashed_document_count
+            TrashedDocument.objects.count(), trashed_document_count
         )
 
     def test_trashed_document_delete_post_view_no_permission(self):
         self.test_document.delete()
         self.assertEqual(Document.valid.count(), 0)
-        self.assertEqual(DeletedDocument.objects.count(), 1)
+        self.assertEqual(TrashedDocument.objects.count(), 1)
 
         response = self._request_test_trashed_document_delete_post_view()
         self.assertEqual(response.status_code, 404)
 
         self.assertEqual(Document.valid.count(), 0)
-        self.assertEqual(DeletedDocument.objects.count(), 1)
+        self.assertEqual(TrashedDocument.objects.count(), 1)
 
     def test_trashed_document_delete_post_view_with_access(self):
         self.test_document.delete()
         self.assertEqual(Document.valid.count(), 0)
-        self.assertEqual(DeletedDocument.objects.count(), 1)
+        self.assertEqual(TrashedDocument.objects.count(), 1)
 
         self.grant_access(
             obj=self.test_document, permission=permission_trashed_document_delete
@@ -155,7 +155,7 @@ class TrashedDocumentViewTestCase(
         response = self._request_test_trashed_document_delete_post_view()
         self.assertEqual(response.status_code, 302)
 
-        self.assertEqual(DeletedDocument.objects.count(), 0)
+        self.assertEqual(TrashedDocument.objects.count(), 0)
         self.assertEqual(Document.valid.count(), 0)
 
     def test_trashed_document_list_view_no_permission(self):
@@ -184,19 +184,19 @@ class TrashCanViewTestCase(
 ):
     def test_trash_can_empty_view_no_permission(self):
         self.test_document.delete()
-        self.assertEqual(DeletedDocument.objects.count(), 1)
+        self.assertEqual(TrashedDocument.objects.count(), 1)
 
         response = self._request_empty_trash_view()
         self.assertEqual(response.status_code, 403)
-        self.assertEqual(DeletedDocument.objects.count(), 1)
+        self.assertEqual(TrashedDocument.objects.count(), 1)
 
     def test_trash_can_empty_view_with_permission(self):
         self.test_document.delete()
-        self.assertEqual(DeletedDocument.objects.count(), 1)
+        self.assertEqual(TrashedDocument.objects.count(), 1)
 
         self.grant_permission(permission=permission_trash_empty)
 
         response = self._request_empty_trash_view()
         self.assertEqual(response.status_code, 302)
-        self.assertEqual(DeletedDocument.objects.count(), 0)
+        self.assertEqual(TrashedDocument.objects.count(), 0)
         self.assertEqual(Document.objects.count(), 0)